♻️ OrderNew.js 입고 단가 입력 시 쉼표 제거

+++ server/highfourm/src/main/java/himedia/project/highfourm/dto/material/MaterialOrderRequestDTO.java
@@ -15,7 +15,7 @@ public class MaterialOrderRequestDTO {
 	private String orderDate;
 	private String standard;
 	private String supplier;
-	private Integer unitPrice;
+	private Long unitPrice;
 	private Long orderAmount;
 	private String note;
